refactor: move condition init test to status_types tests
This makes the tests more concise and sensical as to what is actually
being tested.
